Unicode support/first steps towards gameplay.
An extensive list of changes for this commit:

Isomatrix:
- Full Unicode support has been added, making the game 100%
translatable.  Included the freely available, Unicode-supporting Roboto
font for ingame use.
- Isomatrix is now state-based and contains wrapper classes for your
game and its various states.  See the
com.sigmatauproductions.isomatrix.game package for more information.
- MouseHandler.java has been removed because Slick already handles input
for us just fine.
- Updated Globals.java to contain a linear interpolation function.
- Documented the Prop.createCluster() functionality, and sped it up by
over 250,000% (that's not an exaggeration, either--it was painfully slow
before).
- Tiles now support mouseover actions, and games can request which tile
currently has the mouse over it via TileMap.getMouseOverIndex().
- Added the TileMap.gotoTile() method, which is useful for pathfinding.
- Tilesets now also track their directory name in addition to their
proper tileset.cfg-based name.
- ConfigFile.java now has functions for getting the full filename
(relative to the classpath) of the config file, as well as a string
array containing all of the lines in the file.
- The game now boots based on settings found in
./resources/isomatrix.cfg.

Sigtopolis:
- First build of the game... with an actual game!  No gameplay yet, but
maps are randomly generated with trees and a cursor over the currently
mouseover'd tile.  Humble beginnings.
- A whole slew of GUI options and controls have been added to
com.sigmatauproductions.sigtopolis.ui.  Buttons, checkboxes, dropdown
menus, check it out!
- A FontHandler and WordWrap utility class have been included for
convenience.
- Included (and created) tree graphics.
- Included temporary window icons.